+++ lib/models/active_call.dart
@@ -12,18 +12,16 @@ class ActiveCall {
   // Only available after Ringing and Answer events
   final Map<String, dynamic>? customParams;
 
-  ActiveCall({
-    required String from,
-    required String to,
-    this.initiated,
-    required this.callDirection,
-    this.customParams
-  })   : this.to = to.replaceAll("client:", ""),
+  ActiveCall({required String from, required String to, this.initiated, required this.callDirection, this.customParams})
+      : this.to = to.replaceAll("client:", ""),
         this.from = from.replaceAll("client:", ""),
         toFormatted = _prettyPrintNumber(to),
         fromFormatted = _prettyPrintNumber(from);
 
   static String _prettyPrintNumber(String phoneNumber) {
+    if (phoneNumber.isEmpty) {
+      return "";
+    }
     if (phoneNumber.indexOf('client:') > -1) {
       return phoneNumber.split(':')[1];
     }
@@ -40,11 +38,19 @@ class ActiveCall {
     if (phoneNumber.length == 11) {
       start = 1;
     }
-    return "(" +
-        phoneNumber.substring(start, start + 3) +
-        ") " +
-        phoneNumber.substring(start + 3, start + 6) +
-        "-" +
-        phoneNumber.substring(start + 6);
+    return "(" + phoneNumber.substring(start, start + 3) + ") " + phoneNumber.substring(start + 3, start + 6) + "-" + phoneNumber.substring(start + 6);
+  }
+
+  @override
+  String toString() {
+    return 'ActiveCall{'
+        'to: $to, '
+        'toFormatted: $toFormatted, '
+        'from: $from, '
+        'fromFormatted: $fromFormatted, '
+        'initiated: $initiated, '
+        'callDirection: $callDirection, '
+        'customParams: $customParams, '
+        '}';
   }
 }